move exceptions into castor_api_client
