+++ tests/test_api_endpoints/test_metadata_type_endpoints.py
@@ -9,7 +9,7 @@ https://orcid.org/0000-0003-3052-596X
 import pytest
 
 from tests.test_api_endpoints.data_models import metadata_type_model
-from exceptions.exceptions import CastorException
+from castoredc_api_client.exceptions import CastorException
 
 
 class TestMetadataType:
