+++ tests/test_api_endpoints/test_survey_data_entry_endpoints.py
@@ -8,7 +8,7 @@ https://orcid.org/0000-0003-3052-596X
 """
 import pytest
 
-from exceptions.exceptions import CastorException
+from castoredc_api_client.exceptions import CastorException
 from tests.test_api_endpoints.data_models import (
     survey_data_point_extended_model,
     data_options,
